+++ examples/index.js
@@ -1,17 +1,16 @@
-import Vue from "vue";
-import App from "./App.vue";
-import router from "./router";
-import VeeValidate from "vee-validate";
 import axios from "axios";
-import { MockData } from "./MockData/index.js";
-
+import VeeValidate from "vee-validate";
+import Vue from "vue";
 import VlhForms from "../src/lib";
+import App from "./App.vue";
+import { MockData } from "./MockData/index.js";
+import router from "./router";
 
 const delay = ms => new Promise(resolve => setTimeout(resolve, ms));
 
 const ValidationConfig = {
-  classes: true,
-  events: 'change|blur'
+  classes: true
+  // events: 'change|blur'
 };
 Vue.use(VeeValidate, ValidationConfig);
 Vue.use(VlhForms);
@@ -54,8 +53,8 @@ Vue.mixin({
 
   computed: {
     lastStep() {
-      return this.currentStep === this.totalSteps
-    },    
+      return this.currentStep === this.totalSteps;
+    },
     selectedDegreeLevelObject() {
       return this.getDegreeLevelObject();
     },
@@ -118,7 +117,16 @@ Vue.mixin({
         FirstFormError.focus();
       });
     },
-
+    // stepFormTransitionHandler() {
+    //   if (this.$root.currentStep = 1) {
+    //     return;
+    //   }
+    //   this.$nextTick(function() {
+    //     const TlhForm = document.getElementById('tlh-form');
+    //     const FirstField = TlhForm.querySelector('input, select');
+    //     FirstField.focus()
+    //   });
+    // },
     clearFormOnSubmission() {
       for (let key in this.submit) {
         this.submit[key] = "";
@@ -197,20 +205,11 @@ Vue.mixin({
       this.$bus.$emit("set-focus", "firstName");
     },
 
-    setFocusFN() {
-      this.$bus.$emit("set-focus", "firstName");
-      // this.$refs.formFirstName.$refs.firstName.focus()
-    },
-
-    setFocusLN() {
-      this.$bus.$emit("set-focus", "lastName");
-      // this.$refs.formLastName.$refs.lastName.focus()
-    },
     validateStep() {
       this.$validator.validateAll().then(result => {
-        console.log("Missing fields or errors!");
         if (result) {
           this.$root.currentStep = this.$root.currentStep + 1;
+          this.$root.stepDirection = "positive";
         } else {
           this.setFocusOnFirstFormError();
         }
@@ -218,16 +217,32 @@ Vue.mixin({
     },
 
     handleNextStep() {
-      console.log("stopped");
+      console.log("handleNextStep RAN@!");
       this.validateStep();
-      // let wasCurrentStep = this.$root.currentStep
-      // this.$bus.$emit('next-clicked', { pastStep: wasCurrentStep, newStep: this.$root.currentStep + 1})
-      // this.$root.currentStep = this.$root.currentStep + 1
     },
 
     handlePreviousStep() {
       this.$root.currentStep = this.$root.currentStep - 1;
-      this.$bus.$emit("previous-here");
+      this.$root.stepDirection = "negative";
+    },
+
+    handleFormStepTransition(el) {
+      if (this.$root.currentStep > 1) this.setFocusOnFirstField();
+      // if (this.$root.currentStep === 1 && this.$root.stepDirection === 'negative') this.setFocusOnNextButton()
+    },
+
+    setFocusOnNextButton() {
+      // TODO
+      console.log("SET FOCUS ON NEXT BUTTON HERE");
+    },
+
+    setFocusOnFirstField() {
+      this.$nextTick(() => {
+        const FirstField = document
+          .getElementById("tlh-form")
+          .querySelector("input, select");
+        FirstField.focus();
+      });
     },
 
     registerZipValidator() {