+++ examples/views/Basic.vue
@@ -10,32 +10,23 @@
       id="program"
       :options="programs"
       label="Select a Program "
-      :default-text="submit.program ? '' : 'Select (TEST)'"
       v-model="submit.program"
       @focus="onFocus('program')"
       @change="onChange('program')"
     />
     <button
       @click.prevent="emitQualifier('7cf6c854-ec67-42d7-81eb-4556bc539800');"
-    >
-      Emit Real Online Value
-    </button>
+    >Emit Real Online Value</button>
+    
+    <button @click.prevent="emitQualifier('987654321');">Emit Mock Campus Value</button>
 
-    <button @click.prevent="emitQualifier('987654321');">
-      Emit Mock Campus Value
-    </button>
-
-    <form-first-name v-model="submit.firstName" />
-    <form-last-name v-model="submit.lastName" />
+    <form-first-name v-model="submit.firstName"/>
+    <form-last-name v-model="submit.lastName"/>
     <!-- <form-phone v-model="submit.phone" validation="required"/>
     <form-email v-model="submit.email" validation="required|email"></form-email>
-    <form-zip placeholder="Your Zip" v-model="submit.zip"></form-zip> -->
+    <form-zip placeholder="Your Zip" v-model="submit.zip"></form-zip>-->
     <form-submit :disableOnErrors="true" text="Get New"></form-submit>
-    <form-legal-text
-      school="Alvernia University"
-      fontSize="1.1em"
-      textColor="blue"
-    ></form-legal-text>
+    <form-legal-text school="Alvernia University" fontSize="1.1em" textColor="blue"></form-legal-text>
   </form>
 </template>
 
@@ -50,11 +41,11 @@ export default {
   },
   methods: {
     onFocus(formField) {
-      console.log('focus: ', formField)
+      console.log("focus: ", formField);
     },
 
     onChange(formField) {
-      console.log('change: ', formField)
+      console.log("change: ", formField);
     },
 
     emitQualifier(newValue) {