+++ examples/views/StepForm.vue
@@ -1,68 +1,80 @@
 <template lang="html">
+  <div class="form-wrapper">
+    <h2>Content ABOVE the form</h2>
+    <p>Lorem ipsum dolor, sit amet consectetur adipisicing elit. Optio accusamus eveniet eligendi? Maxime voluptate quis, fuga natus architecto, labore error deleniti sapiente facere iste, modi nobis pariatur quo tempora fugiat ab. Nihil temporibus atque quibusdam, animi itaque fugiat quas odit?</p>
+    <p>Voluptatem, cumque in. Fugiat voluptatum nobis molestias ipsum eius iure, consectetur expedita doloremque voluptatem cumque maiores quidem, libero quasi natus dicta! Voluptas explicabo, minima numquam hic tenetur vero rem magnam amet labore saepe corrupti omnis atque inventore, dignissimos, quia architecto.</p>
+    <p>Libero, expedita repellendus! Qui iure fugit, quae molestias numquam maxime cupiditate impedit ex! Porro omnis a, eaque deserunt quaerat amet tenetur, cupiditate nesciunt quibusdam libero soluta natus perspiciatis, consequatur dolorem atque nulla fuga mollitia pariatur culpa itaque numquam reprehenderit iste.</p>
   <form
     id="tlh-form"
     class="example basic"
     @submit.prevent="handleFormSubmission"
   >
     <div class="stepsWrapper">
-      <!-- <transition name="slide-fade" mode="out-in" appear> -->
-    <transition name="slide-fade" mode="out-in" appear>
-        <form-step v-if="$root.currentStep == 1" :key="1" >
-          <!-- <form-select
-            id="degreeLevel"
-            label="Select Degree Level"
-            v-model="selectedDegreeLevel"
-            :options="levels"
-          />
 
-            <form-select
-              v-if="selectedDegreeLevel"
-              id="program"
-              label="Select a Program"
-              v-model="submit.program"
-              :options="programsForSelectedDegreeLevel"
-            /> -->
-          <form-select
+      <transition name="slide-fade" mode="out-in" appear @after-enter="handleFormStepTransition">
+        <form-step v-if="$root.currentStep == 1" :key="1" >
+
+          <!-- <form-select
             id="program"
             :options="programs"
-            label="Select a Program 1"
+            label="Select a Program"
             v-model="submit.program"
-          />
-          <form-email
-            v-model="submit.email"
-            validation="required|email"
-          /> 
+          /> -->
+          <form-first-name v-model="submit.firstName" />
+          <form-last-name v-model="submit.lastName" />    
+          
+          <form-buttons-wrapper>
+            <next-step-button @next-step="handleNextStep"/>
+          </form-buttons-wrapper>                
         </form-step>
 
         <form-step v-if="$root.currentStep == 2" :key="2">
-          <form-first-name ref="firstField" v-model="submit.firstName" />
-          <form-last-name v-model="submit.lastName" />
+          <form-email v-model="submit.email" validation="required|email"/>
+          <form-phone v-model="submit.phone" validation="required"/>
+          <form-zip v-model="submit.zip" />
+
+          <form-buttons-wrapper>
+            <previous-step-button  />
+            <form-submit />    
+          </form-buttons-wrapper>
+
+          <form-legal-text/>
+        
         </form-step>
       </transition>
     </div>
 
-    <step-form-controls
+      <!-- :key="$root.currentStep" -->
+    <!-- <step-form-controls
       @previous-step="handlePreviousStep"
       @next-step="handleNextStep"
       submitBtnText="BtnText From Prop"
-    />
+    /> -->
     <form-legal-text />
   </form>
+    <h2>Content BELOW the form</h2>
+    <p>Lorem ipsum dolor, sit amet consectetur adipisicing elit. Optio accusamus eveniet eligendi? Maxime voluptate quis, fuga natus architecto, labore error deleniti sapiente facere iste, modi nobis pariatur quo tempora fugiat ab. Nihil temporibus atque quibusdam, animi itaque fugiat quas odit?</p>
+    <p>Voluptatem, cumque in. Fugiat voluptatum nobis molestias ipsum eius iure, consectetur expedita doloremque voluptatem cumque maiores quidem, libero quasi natus dicta! Voluptas explicabo, minima numquam hic tenetur vero rem magnam amet labore saepe corrupti omnis atque inventore, dignissimos, quia architecto.</p>
+    <p>Libero, expedita repellendus! Qui iure fugit, quae molestias numquam maxime cupiditate impedit ex! Porro omnis a, eaque deserunt quaerat amet tenetur, cupiditate nesciunt quibusdam libero soluta natus perspiciatis, consequatur dolorem atque nulla fuga mollitia pariatur culpa itaque numquam reprehenderit iste.</p>  
+  </div>
 </template>
 
 <script>
 export default {
-  mounted () { },
+  mounted() {
+    console.log("MOUNTED: ", this);
+  },
 
   methods: {
-    afterEnter (el) {
-      console.log("el: ", el);
-      // this.$refs.firstField.setFocus();
-
-      // this.$nextTick(() => {
-      //   this.$refs.firstField.setFocus()
-      // })
-    }
+    // afterEnter (el) {
+    //   if (this.$root.currentStep > 1) this.setFocusOnFirstField()
+    // },
+    // setFocusOnFirstField () {
+    //   this.$nextTick(() => {
+    //     const FirstField = document.getElementById('tlh-form').querySelector('input, select');
+    //     FirstField.focus()
+    //   });
+    // }
   }
 };
 </script>