+++ src/mixins/BaseMixin.js
@@ -18,7 +18,7 @@ export default {
      */
     label: {
       type: String,
-      required: true
+      required: false
     },
 
     /**