+++ vue.config.js
@@ -8,6 +8,7 @@ module.exports = {
       libraryExport: "default"
     }
   },
+
   chainWebpack: config => {
     const svgRule = config.module.rule("svg");
     svgRule.uses.clear();